# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OMPay",
        "description": "API pour la gestion des comptes bancaires OMPay",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/comptes/{compteId}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer un compte spécifique",
                "operationId": "363ad13052b7e1e7ec4cd9c96fe29bf7",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "FORBIDDEN"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Vous n'avez pas accès à ce compte"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "compteId": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "COMPTE_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Le compte avec l'ID spécifié n'existe pas"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "compteId": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "C00123456"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Amadou Diallo"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "courant",
                            "epargne",
                            "entreprise",
                            "joint"
                        ],
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1250000
                    },
                    "devise": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-15T00:00:00Z"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "suspendu"
                        ],
                        "example": "bloque"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": "Inactivité de 30+ jours",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-06-10T14:30:00Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}